chore: rename Android package name
