+++ react-native-line.podspec
@@ -3,7 +3,7 @@ require 'json'
 package = JSON.parse(File.read(File.join(__dir__, 'package.json')))
 
 Pod::Spec.new do |s|
-  s.name            = 'RTNLine'
+  s.name            = 'react-native-line'
   s.version         = package['version']
   s.summary         = package['description']
   s.description     = package["description"]
