+++ src/index.ts
@@ -1,4 +1,4 @@
 import NativeLineLogin from './NativeLineLogin'
 
 export * from './NativeLineLogin'
-export default NativeLineLogin!
+export default NativeLineLogin
